# Speed up shader compilation and avoid WebGL crashes