--- a/frontend/javascripts/viewer/shaders/main_data_shaders.glsl.ts
+++ b/frontend/javascripts/viewer/shaders/main_data_shaders.glsl.ts
@@ -169,7 +169,6 @@ precision highp float;
 ${SHARED_UNIFORM_DECLARATIONS}
 
 flat in vec2 index;
-flat in uvec4 outputCompressedEntry[<%= globalLayerCount %>];
 flat in uint outputMagIdx[<%= globalLayerCount %>];
 flat in uint outputSeed[<%= globalLayerCount %>];
 flat in float outputAddress[<%= globalLayerCount %>];
@@ -435,7 +434,6 @@ out mat4 savedModelMatrix;
 }) %>
 
 flat out vec2 index;
-flat out uvec4 outputCompressedEntry[<%= globalLayerCount %>];
 flat out uint outputMagIdx[<%= globalLayerCount %>];
 flat out uint outputSeed[<%= globalLayerCount %>];
 flat out float outputAddress[<%= globalLayerCount %>];