Use loops for bi- and trilinear filtering in shader to speed up shader compilation and avoid WebGL crashes for datasets with many layers.

Turns out, WebGL drivers usually can optimize loops much better than unrolled ones. In this case, the `getColorForCoords` function is very complex which exacerbates the issue.

There's still a lot of room for improvement, but this is a low-hanging fruit we should definitely pick.

Also allow to disable interpolation in flight and oblique mode and recompute the shader based on whether it is enabled or not. The version for disabled interpolation is much less complex and compiles a lot faster.

As a followup, I'll add a fallback that if interpolation is enabled and the shader compilation crashes, it gets disabled and shader compilation will be retried. I'll also add some analytics events, for example time to first render.

### URL of deployed dev instance (used for testing):
- https://avoidwebglcrashes.webknossos.xyz

### Steps to test:
- Open a dataset with more than 3 color layers or a dataset that crashes WebGL on master on your machine (for example, the data_types dataset, at least for my machine). WebGL should no longer crash. Also other datasets should load faster and with less lag than on master.
- Measurements for my machine obtained using the Chrome performance tab:
  - 1-color layer 1.7s GPU freeze (now no freeze)
  - 2-color layer 7s GPU freeze (now 0.9s)
  - 3-color layer 14s GPU freeze (now 1.8s)
  - 4-color layer crash (now 2.4s)
  - 5-color layer crash (now 4.5s)
  - 8-color layer crash (now 10.8s)
  - 1-color 1-segmentation layer 3s GPU freeze (now 0.2s)
- Play around with the interpolation setting in all view modes. There might be a freeze during the first toggle since the shader is recompiled. Afterwards, it's usually cached.
- [x] Run the screenshot tests for this branch ([CI run](https://github.com/scalableminds/webknossos/actions/runs/18188353576))

### Issues:
- fixes WebGL crashes for datasets with many layers

------
(Please delete unneeded items, merge only when none are left open)
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in `unreleased_changes` or use `./tools/create-changelog-entry.py`)
- [x] Removed dev-only changes like prints and application.conf edits
- [x] Considered [common edge cases](../blob/master/.github/common_edge_cases.md)